splitting config into modules
