+++ lua/core/transparency.lua
@@ -1,0 +1,28 @@
+-- ============================================================================
+-- TRANSPARENCY SETUP
+-- ============================================================================
+vim.opt.termguicolors = true
+-- Colorscheme will be set after plugins load
+
+local function set_transparent()
+	local groups = {
+		"Normal",
+		"NormalNC",
+		"EndOfBuffer",
+		"NormalFloat",
+		"FloatBorder",
+		"SignColumn",
+		"StatusLine",
+		"StatusLineNC",
+		"TabLine",
+		"TabLineFill",
+		"TabLineSel",
+		"ColorColumn",
+	}
+	for _, g in ipairs(groups) do
+		vim.api.nvim_set_hl(0, g, { bg = "none" })
+	end
+	vim.api.nvim_set_hl(0, "TabLineFill", { bg = "none", fg = "#767676" })
+end
+
+set_transparent()
