+++ lua/plugins/colorscheme.lua
@@ -1,0 +1,8 @@
+-- ============================================================================
+-- COLORSCHEME SETUP
+-- ============================================================================
+
+require("everforest").setup({
+	-- Your config here
+})
+vim.cmd.colorscheme("everforest")
