+++ lua/plugins/init.lua
@@ -1,0 +1,56 @@
+-- ============================================================================
+-- PLUGIN MANAGEMENT (vim.pack)
+-- ============================================================================
+vim.pack.add({
+	"https://www.github.com/lewis6991/gitsigns.nvim",
+	"https://www.github.com/echasnovski/mini.nvim",
+	"https://www.github.com/ibhagwan/fzf-lua",
+	"https://www.github.com/nvim-tree/nvim-tree.lua",
+	{ src = "https://github.com/neanias/everforest-nvim", lazy = false, priority = 1000 },
+	{
+		src = "https://github.com/nvim-treesitter/nvim-treesitter",
+		branch = "main",
+		build = ":TSUpdate",
+	},
+	-- Language Server Protocols
+	"https://www.github.com/neovim/nvim-lspconfig",
+	"https://github.com/mason-org/mason.nvim",
+	"https://github.com/creativenull/efmls-configs-nvim",
+	{
+		src = "https://github.com/saghen/blink.cmp",
+		version = vim.version.range("1.*"),
+	},
+	"https://github.com/L3MON4D3/LuaSnip",
+})
+
+local function packadd(name)
+	vim.cmd("packadd " .. name)
+end
+packadd("nvim-treesitter")
+packadd("gitsigns.nvim")
+packadd("mini.nvim")
+packadd("fzf-lua")
+packadd("nvim-tree.lua")
+packadd("everforest-nvim")
+-- LSP
+packadd("nvim-lspconfig")
+packadd("mason.nvim")
+packadd("efmls-configs-nvim")
+packadd("blink.cmp")
+packadd("LuaSnip")
+
+-- ============================================================================
+-- PLUGIN CONFIGS
+-- ============================================================================
+
+-- Load plugin configurations in order
+require("plugins.colorscheme")
+require("plugins.treesitter")
+require("plugins.completion")
+require("plugins.lsp")
+require("plugins.nvim-tree")
+require("plugins.fzf-lua")
+require("plugins.mini-plugins")
+require("plugins.gitsigns")
+require("plugins.terminal")
+require("plugins.claude")
