+++ lua/plugins/lsp/init.lua
@@ -1,0 +1,9 @@
+-- ============================================================================
+-- LSP MODULE ORCHESTRATION
+-- ============================================================================
+
+-- Load LSP modules in correct order
+require("plugins.lsp.config")
+require("plugins.lsp.handlers")
+require("plugins.lsp.efm")
+require("plugins.lsp.servers")
