+++ lua/plugins/mini-plugins.lua
@@ -1,0 +1,15 @@
+-- ============================================================================
+-- MINI.NVIM PLUGINS CONFIG
+-- ============================================================================
+
+require("mini.ai").setup({})
+require("mini.comment").setup({})
+require("mini.move").setup({})
+require("mini.surround").setup({})
+require("mini.cursorword").setup({})
+require("mini.indentscope").setup({})
+require("mini.pairs").setup({})
+require("mini.trailspace").setup({})
+require("mini.bufremove").setup({})
+require("mini.notify").setup({})
+require("mini.icons").setup({})
